ignoreip patch for djbdns
Ignore the IP address in $IGNOREIP

This patch changes djbdns's dnscache program so that it will ignore the
IP address given in the IGNOREIP environment variable.  I wrote this
patch because of Verisign's oh-so helpful wildcard A records for *.COM
and *.NET.  Install the patch like this:
    patch <djbdns-1.05-ignoreip.patch
    svc -d /service/dnscache
    make setup check
    svc -u /service/dnscache

Configure it to ignore Verisign's wildcard record like this:
    echo 64.94.110.11 >/service/dnscache/env/IGNOREIP
    svc -t /service/dnscache

If IGNOREIP isn't found in the environment variable space, nothing is
ignored.

--
--My blog is at angry-economist.russnelson.com  | Free markets express in the
Crynwr sells support for free software  | PGPok | practical world our belief
521 Pleasant Valley Rd. | +1 315 268 1925 voice | that there is that of God
Potsdam, NY 13676-3213  | +1 315 268 9201 FAX   | in all people. -Chris V.