Reuse the socket when resending the same query
Previously, dnscache would create a new socket for each UDP request sent
to a name server. If one query required several requests to different
servers because, say, the server responses timed out, dnscache would
close and reopen the socket each time.

Now, each time a request is resent to a different nameserver, reuse
the socket.  Don't bind the socket to a destination, and use sendto()
instead of send(). Process the first response received. The socket is
subsequently closed, so any later responses will be ignored.

In a high latency situation (e.g. 5 second RTT on responses) this
results in a considerable improvement in speed and reliability. A
query for 'A www.telstra.net' took 310 sec in unpatched dnscache
and 16 sec after this change was applied.

Signed-off-by: Nick Andrew <nick@nick-andrew.net>